添加直播发起界面、修改界面

+++ M8Tool-master/M8Tool/M8Tool/ViewControllers/Meeting/MeetingLuanch/MeetingMembersCollection.m
@@ -216,7 +216,7 @@ static NSString *CollectionHeaderID = @"MeetingMembersCollectionHeaderID";
     else
     {
         [AppDelegate showAlertWithTitle:@"温馨提示"
-                                message:[NSString stringWithFormat:@"最多只能邀请: %ld 人", self.totalNumbers]
+                                message:[NSString stringWithFormat:@"最多只能邀请: %ld 人", (long)self.totalNumbers]
                                 okTitle:@"确定"
                             cancelTitle:nil
                                      ok:nil
